Unify class attendance page with enriched attendance table
