+++ web/app/routes/manage/attendance.tsx
@@ -1,3 +1,7 @@
+import { requireAuth } from '@/lib/auth.server'
+import { Constants, type Database } from '@/lib/database.types'
+import { isRoleAtLeast } from '@/lib/roles'
+import { adminClient } from '@/lib/supabase/adminClient'
 import { createClient } from '@/lib/supabase/server'
 import type { Route } from './+types/attendance'
 import TableDisplay from './table-display'
@@ -73,6 +77,11 @@ const displayName = (profile: ProfileRow | null) => {
   return ''
 }
 
+const displayNameOrId = (profile: ProfileRow | null, fallbackId: string) => {
+  const label = displayName(profile)
+  return label || `ID ${fallbackId}`
+}
+
 const statusLabel = ({ done, total }: { done: number; total: number }) => {
   if (total <= 0) return 'N/A'
   if (done >= total) return `Done (${done}/${total})`
@@ -81,8 +90,8 @@ const statusLabel = ({ done, total }: { done: number; total: number }) => {
 }
 
 export async function loader({ request }: Route.LoaderArgs) {
-  const { supabase } = createClient(request)
-  const { data, error } = await supabase
+  const auth = await requireAuth(request)
+  const { data, error } = await adminClient
     .from('class_attendance')
     .select('id, class_id, profile_id, status, photo_status, camera_on, recorded_by, created_at, updated_at')
     .order('created_at', { ascending: false })
@@ -94,22 +103,25 @@ export async function loader({ request }: Route.LoaderArgs) {
   const attendanceRows = (data ?? []) as AttendanceRow[]
   const classIds = Array.from(new Set(attendanceRows.map(row => row.class_id).filter(Boolean)))
   const profileIds = Array.from(new Set(attendanceRows.map(row => row.profile_id).filter(Boolean)))
+  const recordedByIds = Array.from(new Set(attendanceRows.map(row => row.recorded_by).filter((id): id is string => Boolean(id))))
+
+  const profileLookupIds = Array.from(new Set([...profileIds, ...recordedByIds]))
 
   const [{ data: classRows }, { data: profileRows }, { data: meetingRows }, { data: registrantRows }] = await Promise.all([
     classIds.length
-      ? supabase.from('class').select('id, workshop_id, starts_at, ends_at').in('id', classIds)
+      ? adminClient.from('class').select('id, workshop_id, starts_at, ends_at').in('id', classIds)
       : Promise.resolve({ data: [] as ClassRow[] }),
-    profileIds.length
-      ? supabase.from('profile').select('id, firstname, surname, email').in('id', profileIds)
+    profileLookupIds.length
+      ? adminClient.from('profile').select('id, firstname, surname, email, user_id').in('id', profileLookupIds)
       : Promise.resolve({ data: [] as ProfileRow[] }),
     classIds.length
-      ? supabase
+      ? adminClient
           .from('class_zoom_meeting')
           .select('id, class_id, status, zoom_meeting_id, topic, start_time, duration_minutes, join_url, host_zoom_user_email')
           .in('class_id', classIds)
       : Promise.resolve({ data: [] as MeetingRow[] }),
     classIds.length
-      ? supabase
+      ? adminClient
           .from('class_zoom_registrant')
           .select('class_id, profile_id, zoom_registrant_id, zoom_join_url, last_sent_at')
           .in('class_id', classIds)
@@ -122,10 +134,10 @@ export async function loader({ request }: Route.LoaderArgs) {
 
   const [{ data: workshopRows }, { data: enrollmentRows }, { data: syncRows }] = await Promise.all([
     workshopsIds.length
-      ? supabase.from('workshop').select('id, description').in('id', workshopsIds)
+      ? adminClient.from('workshop').select('id, description').in('id', workshopsIds)
       : Promise.resolve({ data: [] as WorkshopRow[] }),
     workshopsIds.length
-      ? supabase
+      ? adminClient
           .from('workshop_enrollment')
           .select('workshop_id, profile_id')
           .in('workshop_id', workshopsIds)
@@ -133,7 +145,7 @@ export async function loader({ request }: Route.LoaderArgs) {
           .not('profile_id', 'is', null)
       : Promise.resolve({ data: [] as EnrollmentRow[] }),
     meetingIds.length
-      ? supabase
+      ? adminClient
           .from('class_zoom_participant_sync')
           .select('class_zoom_meeting_id, status, created_at')
           .in('class_zoom_meeting_id', meetingIds)
@@ -143,7 +155,13 @@ export async function loader({ request }: Route.LoaderArgs) {
 
   const classById = new Map(classes.map(row => [row.id, row]))
   const workshopById = new Map(((workshopRows ?? []) as WorkshopRow[]).map(row => [row.id, row]))
-  const profileById = new Map(((profileRows ?? []) as ProfileRow[]).map(row => [row.id, row]))
+  const profileRowsTyped = (profileRows ?? []) as Array<ProfileRow & { user_id?: string | null }>
+  const profileById = new Map(profileRowsTyped.map(row => [row.id, row]))
+  const profileByUserId = new Map(
+    profileRowsTyped
+      .map(row => (typeof row.user_id === 'string' && row.user_id ? [row.user_id, row] : null))
+      .filter((entry): entry is [string, ProfileRow & { user_id?: string | null }] => Boolean(entry))
+  )
 
   const meetingByClassId = new Map<string, MeetingRow>()
   for (const row of (meetingRows ?? []) as MeetingRow[]) {
@@ -222,7 +240,7 @@ export async function loader({ request }: Route.LoaderArgs) {
       workshop_description: workshop?.description ?? 'Workshop',
       class_starts_at: classRow?.starts_at ?? null,
       class_ends_at: classRow?.ends_at ?? null,
-      profile_display: displayName(profile),
+      profile_display: displayNameOrId(profile, row.profile_id),
       student_join_url: studentRegistrant?.zoom_join_url ?? null,
       zoom_meeting_id: meeting?.zoom_meeting_id ?? null,
       zoom_topic: meeting?.topic ?? null,
@@ -235,6 +253,10 @@ export async function loader({ request }: Route.LoaderArgs) {
       step_attendance_rows: statusLabel({ done: attendanceRowsReady, total: approvedProfiles.size }),
       step_reminder: statusLabel({ done: remindersSent, total: approvedProfiles.size }),
       step_attendance_sync: stepAttendanceSync,
+      recorded_by_email:
+        typeof row.recorded_by === 'string' && row.recorded_by
+          ? displayName(profileByUserId.get(row.recorded_by) ?? null) || row.recorded_by
+          : null,
     }
   })
 
@@ -255,7 +277,7 @@ export async function loader({ request }: Route.LoaderArgs) {
 
   return {
     label: 'Attendance',
-    tableName: 'attendance-raw',
+    tableName: 'class-attendance',
     columns: [
       'workshop_description',
       'class_starts_at',
@@ -276,7 +298,7 @@ export async function loader({ request }: Route.LoaderArgs) {
       'step_attendance_rows',
       'step_reminder',
       'step_attendance_sync',
-      'recorded_by',
+      'recorded_by_email',
       'created_at',
       'updated_at',
       'class_id',
@@ -304,14 +326,86 @@ export async function loader({ request }: Route.LoaderArgs) {
       step_attendance_rows: { label: 'Step 3: Attendance Rows', filterable: true },
       step_reminder: { label: 'Step 4: Reminder', filterable: true },
       step_attendance_sync: { label: 'Step 5: Attendance Sync', filterable: true },
-      recorded_by: { label: 'Recorded by user_id', filterable: true, truncate: true },
+      recorded_by_email: { label: 'Recorded by', filterable: true, truncate: true },
       created_at: { label: 'Created', filterable: true },
       updated_at: { label: 'Updated', filterable: true },
       class_id: { label: 'Class ID', filterable: true },
       profile_id: { label: 'Profile ID', filterable: true },
       id: { label: 'Attendance ID', filterable: true },
     },
+    canEditStatus: isRoleAtLeast(auth.claims.role, 'staff'),
   }
+}
+
+export async function action({ request }: Route.ActionArgs) {
+  const auth = await requireAuth(request)
+  if (!isRoleAtLeast(auth.claims.role, 'staff')) {
+    return new Response('Unauthorized', { status: 403, headers: auth.headers })
+  }
+
+  const formData = await request.formData()
+  const intent = formData.get('intent') as string | null
+  if (intent !== 'update-status' && intent !== 'update-photo-status' && intent !== 'update-camera-on') {
+    return new Response('Unsupported action', { status: 400, headers: auth.headers })
+  }
+
+  const classId = formData.get('class_id') as string
+  const profileId = formData.get('profile_id') as string
+  if (!classId || !profileId) {
+    return new Response('Missing identifiers', { status: 400, headers: auth.headers })
+  }
+
+  const updates: {
+    status?: string | null
+    photo_status?: string | null
+    camera_on?: boolean | null
+    recorded_by: string
+  } = {
+    recorded_by: auth.user.id,
+  }
+
+  if (intent === 'update-status') {
+    const status = (formData.get('status') as string | null) ?? null
+    const allowedStatuses = Constants.public.Enums.class_attendance_status as readonly Database['public']['Enums']['class_attendance_status'][]
+    if (status && !allowedStatuses.includes(status as Database['public']['Enums']['class_attendance_status'])) {
+      return new Response('Invalid status', { status: 400, headers: auth.headers })
+    }
+    updates.status = status || null
+  }
+
+  if (intent === 'update-photo-status') {
+    const photoStatus = (formData.get('photo_status') as string | null) ?? null
+    const allowedPhotoStatuses =
+      Constants.public.Enums.class_attendance_photo_status as readonly Database['public']['Enums']['class_attendance_photo_status'][]
+    if (photoStatus && !allowedPhotoStatuses.includes(photoStatus as Database['public']['Enums']['class_attendance_photo_status'])) {
+      return new Response('Invalid photo status', { status: 400, headers: auth.headers })
+    }
+    updates.photo_status = photoStatus || null
+  }
+
+  if (intent === 'update-camera-on') {
+    const rawCameraOn = (formData.get('camera_on') as string | null) ?? ''
+    if (rawCameraOn === 'true') {
+      updates.camera_on = true
+    } else if (rawCameraOn === 'false') {
+      updates.camera_on = false
+    } else {
+      updates.camera_on = null
+    }
+  }
+
+  const { supabase } = createClient(request)
+  const { error } = await supabase
+    .from('class_attendance')
+    .update(updates)
+    .eq('class_id', classId)
+    .eq('profile_id', profileId)
+
+  if (error) {
+    return new Response(error.message, { status: 500, headers: auth.headers })
+  }
+
+  return { ok: true }
 }
 
 export default function AttendanceRawTablePage() {
